rename flashtarget module to flash_target
